+++ source/packages/services/installer/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @cdf/installer
 
-This log was last generated on Tue, 15 Mar 2022 18:37:51 GMT and should not be manually modified.
+This log was last generated on Thu, 17 Mar 2022 21:52:07 GMT and should not be manually modified.
+
+## 0.1.8
+Thu, 17 Mar 2022 21:52:07 GMT
+
+### Patches
+
+- Fixes issue where installer creates new KMS key on every run
+
+## 0.1.7
+Thu, 17 Mar 2022 05:13:05 GMT
+
+### Patches
+
+- Fix dependency between apigw.cloudFormationTemplate and apigw.type default value
 
 ## 0.1.6
 Tue, 15 Mar 2022 18:37:51 GMT